Update expected_value_dice.py
I made the code a lot more efficient. There were many unnecessary lines I had in the original code while I was starting out, such as defining the number of faces as 0 before asking for the user input. Also added a table to show outcomes and a standard deviation statistic.